gir: generate better output for `document_new_file`
`g-ir-scanner` treats anything with `_new` in the name as a constructor
which, strictly speaking, `document_new_file` is not. Rename this
function to prevent `g-ir-scanner` from seeing it, and then rename
it back in the GIR XML file so that the correct typelib is compiled and
the correct name is in the installed GIR file. This hack is already used
for other symbols.

The result is a static function `Document.new_file` which seems to
be reasonable in languages such as Vala and Python.