doxygen-helper: make use of scintilla indicators
Indicators allow to jump to insert positions and give a visual indication.
Experimental code is included that calls Geany's built-in snippet support,
to be enabled once it's improved to use scintilla indicators as well
(see https://github.com/geany/geany/pull/1470).

Currently a separate keybinding is necessary to make the jump to insertion
points.