doxygen-helper: Initial preview and template editing
The preview is generated from the template and attepmts to use Geany's filetype
styling as much as possible. The template is filled into the edit pane but
you can't save or refresh the preview yet.

Also fixed bugs with the sci indicators which have been off.

Everything is still mostly incomplete.