build-sys: rework vala generator
The vala generator was not very robust. The generateed C files did
not properly depend on the vapi files in api/, which led to build failures.

The combination with the fast_vapi generator didn't really allow
to properly describe this dependency through DEPS or anything.

Also, all vala files were considered to belong together (to build one
library) but that isn't necessarily the case (though within peasy it is).

For the new generator, vala files are grouped explicitly and vapi dependencies
in api/ can be specified normally through DEPS.

Also, for typelib files, the include path to dependency .gir files
couldn't be specified  Use INTROSPECTION_COMPILER_ARGS flag for that.