peasy: install typelibs and PyGI overrides into private directory
This allows for an easy clean removal of peasy because we don't install
into other packages anymore. And, arguably more importantly, it finally
allows to install the PyGI overrides to a version-less directory,
so that an upgrade of the interpreter doesn't necessarily break
peasy plugins anymore.

See https://bugzilla.gnome.org/show_bug.cgi?id=680913 for the pkgutil trick.